requirements.txt
django
djangorestframework
django-cors-headers
django-environ
dj-rest-auth
django-allauth
requests